requirements.txt
flask
sentence-transformers
scikit-learn
pdfminer.six
python-docx
pandas
werkzeug
transformers
